requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
matplotlib>=3.4.0
wandb>=0.12.0
pyyaml>=5.4.0
scipy>=1.7.0
tracywidom>=0.1.0